+++ src/main/java/shortener/domain/urlencoder/algorithm/AdlerHash.java
@@ -1,4 +1,4 @@
-package shortener.infrastructure.urlencoder.algorithm;
+package shortener.domain.urlencoder.algorithm;
 
 public class AdlerHash {
